+++ kompass/kompass/components/component.py
@@ -4,10 +4,11 @@ from omegaconf import OmegaConf
 from ros_sugar.core import ComponentFallbacks, BaseComponent
 from ros_sugar.tf import TFListener, TFListenerConfig
 from ros_sugar.supported_types import add_additional_datatypes
+from ros_sugar.io import Publisher, get_all_msg_types, AllowedTopics
 
 from ..callbacks import GenericCallback
 from ..config import ComponentConfig, RobotConfig, ComponentRunType
-from ..topic import Topic, update_topics, get_all_msg_types, AllowedTopics, Publisher
+from .ros import Topic, update_topics
 from .. import data_types
 from itertools import groupby
